// packages/preset_react/src/schemas/api-schema.ts
export const podcastSearchResultSchema = z.object({
  trackName: z.string(),
  artistName: z.string(),
  artworkUrl100: z.string().optional(),
  feedUrl: z.string().optional(),
  trackCount: z.number().optional(),
  primaryGenreName: z.string().optional(),
});
// ...
export type PodcastSearchResult = z.infer<typeof podcastSearchResultSchema>;